Refactoring code components file created db controllers and queries added to components trouble shooting to refactor code
