+++ app.js
@@ -1,30 +1,30 @@
-const express = require('express');
-const cors = require('cors');
-const db = require('./db');
-const usersController = require('./controllers/usersController');
+const express = require("express");
+const cors = require("cors");
+const db = require("./components/db");
+const usersController = require("./components/controllers/usersController");
 
 const app = express();
 app.use(cors());
 app.use(express.json());
 
-app.get('/', (req, res) => {
-  res.status(200).json({ data: "Service is running "});
+app.get("/", (req, res) => {
+  res.status(200).json({ data: "Service is running " });
 });
 
-app.get('/tests', async (request, response) => {
+app.get("/users", async (request, response) => {
   try {
-    const tests = await db.any('SELECT * FROM tests');
+    const users = await db.any("SELECT * FROM users");
 
-    response.status(200).json({ data: tests });
+    response.status(200).json({ data: users });
   } catch (err) {
     response.status(500).json({ error: err.message });
   }
 });
 
-app.use('/users', usersController);
+app.use("/users", usersController);
 
-app.all('*', (req, res) => {
-  res.status(404).json({ error: "Not found"});
-})
+app.all("*", (req, res) => {
+  res.status(404).json({ error: "Not found" });
+});
 
 module.exports = app;
