+++ components/db/seeds.sql
@@ -1,0 +1,5 @@
+INSERT INTO users (age, name, company, country, photo, about, latitude, longitude)
+    VALUES
+    (35,'Park Daugherty','BUZZWORKS','Greenland','https://randomuser.me/api/portraits/lego/0.jpg', 'Hi there! I''m Park Daugherty, a 35-year-old from Greenland. I work at BUZZWORKS, where I''m passionate about what I do and always strive to exceed expectations. In my free time, I love to spend time with my loved ones and take on new adventures. I''m excited to see what the future holds and what opportunities lie ahead.',38.680966, -70.369272),
+    (29,'Meagan Ashley','ASSISTIX','Romania','https://randomuser.me/api/portraits/lego/1.jpg','Meagan is currently employed at ASSISTIX, where she has quickly established herself as a valuable member of the team. Meagan''s strong work ethic, attention to detail, and creative problem-solving skills have made her a sought-after colleague. Outside of work, Meagan enjoys exploring her city, trying new foods, and spending time with her friends and family. She is a well-rounded individual who is always looking for new opportunities to grow and learn.',39.359043, -140.795149),
+    (29,'Denise Garrison','VURBO','Nepal','https://randomuser.me/api/portraits/lego/2.jpg','I work at Vurbo, where I get to put my skills to the test every day and help drive the company''s success. Outside of work, I love to immerse myself in the local culture and take in all that Nepal has to offer. I''m also an avid traveler and enjoy experiencing new sights, sounds, and tastes.',32.675183, -32.736304 );
