+++ components/queries/usersQueries.js
@@ -1,8 +1,8 @@
 // const usersData = require('../usersData.json');
 
-const db = require('../../db')
+const db = require("../db");
 
-// const idToUserIndex = usersData.reduce((index, user) => {
+// const idToUserIndex = db.reduce((index, user) => {
 //   const { id } = user;
 //   index[id] = user;
 //   return index;
@@ -10,19 +10,16 @@ const db = require('../../db')
 
 // const getAllUsers = () => usersData;
 const getAllUsers = () => {
-  return db.any('SELECT * FROM users');
+  return db.any("SELECT * FROM users");
 };
-
-
-// const getUserById = async (id) => {
-//   return idToUserIndex[id];
-// }
 
 const getUserById = async (id) => {
-  return db.oneOrNone('SELECT * FROM users WHERE id = $1', id);
-};
+//  idToUserIndex[id];
+  return db.oneOrNone("SELECT * FROM users WHERE id = $1", id);
+}
+
 
 module.exports = {
   getAllUsers,
   getUserById,
-}
+};
