--- usersData.json
@@ -1,112 +1,0 @@
-[
-  {
-    "id": 1,
-    "age": 35,
-    "name": "Park Daugherty",
-    "company": "BUZZWORKS",
-    "country": "Greenland",
-    "photo": "https://randomuser.me/api/portraits/lego/0.jpg",
-    "about": "Hi there! I'm Park Daugherty, a 35-year-old from Greenland. I work at BUZZWORKS, where I'm passionate about what I do and always strive to exceed expectations. In my free time, I love to spend time with my loved ones and take on new adventures. I'm excited to see what the future holds and what opportunities lie ahead.",
-    "latitude": 38.680966,
-    "longitude": -70.369272
-  },
-  {
-    "id": 2,
-    "age": 29,
-    "name": "Meagan Ashley",
-    "company": "ASSISTIX",
-    "country": "Romania",
-    "photo": "https://randomuser.me/api/portraits/lego/1.jpg",
-    "about": "Meagan is currently employed at ASSISTIX, where she has quickly established herself as a valuable member of the team. Meagan's strong work ethic, attention to detail, and creative problem-solving skills have made her a sought-after colleague. Outside of work, Meagan enjoys exploring her city, trying new foods, and spending time with her friends and family. She is a well-rounded individual who is always looking for new opportunities to grow and learn.",
-    "latitude": 39.359043,
-    "longitude": -140.795149
-  },
-  {
-    "id": 3,
-    "age": 29,
-    "name": "Denise Garrison",
-    "company": "VURBO",
-    "country": "Nepal",
-    "photo": "https://randomuser.me/api/portraits/lego/2.jpg",
-    "about": "I work at Vurbo, where I get to put my skills to the test every day and help drive the company's success. Outside of work, I love to immerse myself in the local culture and take in all that Nepal has to offer. I'm also an avid traveler and enjoy experiencing new sights, sounds, and tastes.",
-    "latitude": 32.675183,
-    "longitude": -32.736304
-  },
-  {
-    "id": 4,
-    "age": 24,
-    "name": "Carroll Melton",
-    "company": "TURNABOUT",
-    "country": "Bulgaria",
-    "photo": "https://randomuser.me/api/portraits/lego/3.jpg",
-    "about": "Eu do dolor occaecat nulla quis qui. Sint amet dolor nostrud est minim velit aute et. Cupidatat reprehenderit fugiat commodo laboris tempor eu. Reprehenderit incididunt enim nulla qui.\r\n",
-    "latitude": -5.083,
-    "longitude": -155.696156
-  },
-  {
-    "id": 5,
-    "age": 28,
-    "name": "Brigitte Ortega",
-    "company": "INSURITY",
-    "country": "Niue",
-    "photo": "https://randomuser.me/api/portraits/lego/4.jpg",
-    "about": "Ex fugiat mollit nisi consequat nostrud mollit dolor ad officia culpa eiusmod reprehenderit ullamco. Velit eu consequat dolore cupidatat excepteur esse qui pariatur adipisicing reprehenderit sunt nisi in. Veniam excepteur cillum proident nostrud minim aliqua ut velit. Duis enim excepteur proident ullamco sint non culpa sunt sit duis. Cupidatat duis veniam enim sint pariatur occaecat id.\r\n",
-    "latitude": 72.881335,
-    "longitude": -147.668845
-  },
-  {
-    "id": 6,
-    "age": 32,
-    "name": "Keisha Clements",
-    "company": "GEOFORM",
-    "country": "Tuvalu",
-    "photo": "https://randomuser.me/api/portraits/lego/5.jpg",
-    "about": "Sint est incididunt deserunt occaecat quis elit enim in labore. Ipsum dolore ipsum labore sit qui velit adipisicing. Ipsum mollit qui Lorem dolor exercitation excepteur nostrud tempor. Tempor incididunt ipsum anim cupidatat do amet adipisicing ea tempor labore incididunt labore. Velit officia enim amet cillum veniam consequat laboris ea anim pariatur. Sint ex aliquip cillum cillum dolor.\r\n",
-    "latitude": 83.304345,
-    "longitude": -104.973509
-  },
-  {
-    "id": 7,
-    "age": 24,
-    "name": "Rogers Wiley",
-    "company": "POOCHIES",
-    "country": "Senegal",
-    "photo": "https://randomuser.me/api/portraits/lego/6.jpg",
-    "about": "Quis sunt irure exercitation veniam aute et proident reprehenderit irure adipisicing esse ad commodo. Anim enim elit dolor excepteur labore id cillum est. Eiusmod laboris adipisicing ad irure enim do sunt. Adipisicing veniam consequat ad sit ut aliquip.\r\n",
-    "latitude": 0.040734,
-    "longitude": -165.031747
-  },
-  {
-    "id": 8,
-    "age": 21,
-    "name": "Bernard Cruz",
-    "company": "TRANSLINK",
-    "country": "Mozambique",
-    "photo": "https://randomuser.me/api/portraits/lego/7.jpg",
-    "about": "Elit qui velit adipisicing sit ut sunt. Officia aliqua ex irure eu magna nisi minim pariatur proident minim aliquip minim. Id elit consequat nulla et ipsum do pariatur irure. Nisi consectetur mollit ut sit aliqua ea duis enim ullamco sit exercitation veniam.\r\n",
-    "latitude": -10.959657,
-    "longitude": -9.708998
-  },
-  {
-    "id": 9,
-    "age": 33,
-    "name": "Wong Pate",
-    "company": "KLUGGER",
-    "country": "Tajikistan",
-    "photo": "https://randomuser.me/api/portraits/lego/8.jpg",
-    "about": "Dolor sit nulla ipsum laboris ut tempor minim ad consectetur proident velit nisi occaecat. Laborum commodo veniam aliquip non laboris do fugiat in anim sunt veniam aliqua deserunt. Ipsum fugiat cupidatat laborum mollit nisi veniam cupidatat laborum. Eiusmod fugiat culpa tempor irure do. Labore culpa anim veniam aliquip nostrud in non deserunt minim adipisicing et ad ipsum. Consequat nisi laborum qui id magna ipsum labore. Ipsum Lorem non dolore ea anim ex in mollit mollit qui.\r\n",
-    "latitude": 19.605271,
-    "longitude": -106.281246
-  },
-  {
-    "id": 10,
-    "age": 30,
-    "name": "Felecia Meadows",
-    "company": "LOVEPAD",
-    "country": "El Salvador",
-    "photo": "https://randomuser.me/api/portraits/lego/9.jpg",
-    "about": "Voluptate cillum culpa labore ut. Labore nulla consectetur fugiat deserunt excepteur velit deserunt duis proident incididunt voluptate et. Proident qui qui do excepteur consectetur nostrud id nulla excepteur commodo veniam mollit dolor adipisicing. Consequat exercitation enim sint mollit nostrud consequat dolore pariatur magna pariatur. Laboris excepteur voluptate eu ipsum sunt esse dolor elit amet et sunt minim.\r\n",
-    "latitude": 83.252308,
-    "longitude": -124.303889
-  }
-]
